Move ClauseIF to clause.rs; ClauseDB and ClauseDBIF to db.rs
